+++ src/Gateway.php
@@ -7,7 +7,7 @@
  * Company: Pronamic
  *
  * @author Reüel van der Steege
- * @version 1.0.2
+ * @version 1.0.3
  * @since 1.0.0
  */
 class Pronamic_WP_Pay_Extensions_Give_Gateway {
@@ -50,9 +50,15 @@ class Pronamic_WP_Pay_Extensions_Give_Gateway {
 	 * @since   1.0.0
 	 */
 	public function gateway_settings( $settings ) {
+		$description = '';
+
+		if ( 'pronamic_pay' === $this->id ) {
+			$description = __( "This payment method does not use a predefined payment method for the payment. Some payment providers list all activated payment methods for your account to choose from. Use payment method specific gateways (such as 'iDEAL') to let customers choose their desired payment method at checkout.", 'pronamic_ideal' );
+		}
+
 		$settings[] = array(
 			'name' => $this->name,
-			'desc' => '',
+			'desc' => $description,
 			'id'   => sprintf( 'give_title_%s', $this->id ),
 			'type' => 'give_title',
 		);
@@ -64,6 +70,14 @@ class Pronamic_WP_Pay_Extensions_Give_Gateway {
 			'type'    => 'select',
 			'options' => Pronamic_WP_Pay_Plugin::get_config_select_options( $this->payment_method ),
 			'default' => get_option( 'pronamic_pay_config_id' ),
+		);
+
+		$settings[] = array(
+			'name'    => __( 'Transaction description', 'pronamic_ideal' ),
+			'desc'    => sprintf( __( 'Available tags: %s', 'pronamic_ideal' ), sprintf( '<code>%s</code>', '{donation_id}' ) ),
+			'id'      => sprintf( 'give_%s_transaction_description', $this->id ),
+			'type'    => 'text',
+			'default' => __( 'Give donation {donation_id}', 'pronamic_ideal' ),
 		);
 
 		return $settings;
@@ -153,7 +167,7 @@ class Pronamic_WP_Pay_Extensions_Give_Gateway {
 
 			if ( $gateway ) {
 				// Data
-				$data = new Pronamic_WP_Pay_Extensions_Give_PaymentData( $donation_id );
+				$data = new Pronamic_WP_Pay_Extensions_Give_PaymentData( $donation_id, $this );
 
 				$gateway->set_payment_method( $this->payment_method );
 
@@ -183,5 +197,15 @@ class Pronamic_WP_Pay_Extensions_Give_Gateway {
 				}
 			}
 		}
+	}
+
+	/**
+	 * Get transaction description setting.
+	 *
+	 * @since 1.0.3
+	 * @return string
+	 */
+	public function get_transaction_description() {
+		return give_get_option( sprintf( 'give_%s_transaction_description', $this->id ) );
 	}
 }