+++ src/MisterCashGateway.php
@@ -7,7 +7,7 @@
  * Company: Pronamic
  *
  * @author Reüel van der Steege
- * @version 1.0.1
+ * @version 1.0.3
  * @since 1.0.0
  */
 class Pronamic_WP_Pay_Extensions_Give_MisterCashGateway extends Pronamic_WP_Pay_Extensions_Give_Gateway {
@@ -17,8 +17,8 @@ class Pronamic_WP_Pay_Extensions_Give_MisterCashGateway extends Pronamic_WP_Pay_
 	public function __construct() {
 		parent::__construct(
 			'pronamic_pay_mister_cash',
-			__( 'Bancontact/Mister Cash', 'pronamic_ideal' ),
-			Pronamic_WP_Pay_PaymentMethods::MISTER_CASH
+			__( 'Bancontact', 'pronamic_ideal' ),
+			Pronamic_WP_Pay_PaymentMethods::BANCONTACT
 		);
 	}
 }